BUGFIX: Fixed no a.out executable support on windows
On windows, the g++ compiler generates a.exe executable when no output name is
specified. And on linux it's a.out. Using a custom name won't fix the issue
because in windows there would be a .exe suffix to the executable and the
command to execute would also be different.

Hence, detecting the os and changing the execute command if on windows.